-Fin de Tarea Semana6 Manejo de Excepciones

+++ app-web/src/main/java/clase19practicasemanal/InMemoryDB.java
@@ -1,4 +1,5 @@
 package clase19practicasemanal;
+
 /*
 Hola chicos, buenas noches.
 Tal como les comenté en clases, les dejo el enunciado del ejercicio semanal.
@@ -8,38 +9,73 @@ Crear un branch con el nombre practica-adicional-semana6
 - modificar la interface IAction para poder lanzar dichas exceptions
 - modificar la clase MainMap para controlar e informar los errores.
 */
+
 import java.util.HashMap;
 import java.util.Map;
 
+import clase19practicasemanal.exceptions.MemoryDBCRUDException;
+import clase19practicasemanal.exceptions.MemoryDBListarException;
+import clase19practicasemanal.exceptions.MemoryDBModificarException;
+import clase19practicasemanal.exceptions.MemoryDBSaveException;
+import clase19practicasemanal.exceptions.MemoryDBUpdateException;
 import clase7.Articulo;
 
 public class InMemoryDB {
 
-	static private Map<Long,Articulo> db = new HashMap<>();
-	
-	public static void save(Long key,Articulo valor) throws MemoryDBException{
+	static private Map<Long, Articulo> db = new HashMap<>();
+
+	public static void save(Long key, Articulo valor) throws MemoryDBCRUDException {
+		if (key != null && valor != null && valor instanceof Articulo) {
 			InMemoryDB.db.put(key, valor);
-		
-	}
-	
-	public static void update(Long key, Articulo newValue) throws MemoryDBException{
-	
-		InMemoryDB.db.replace(key,newValue);
-	}
-	
-	public static void delete(Long key) throws MemoryDBException{
-		InMemoryDB.db.remove(key);
-	}
-	
-	public static Articulo getById(Long key) throws MemoryDBException{
-		Articulo value = null;
-		if(InMemoryDB.db.containsKey(key)) {
-			value = InMemoryDB.db.get(key);
+		} else if (key == null || valor == null) {
+			throw new MemoryDBCRUDException("Error al Guardar en la Base de Datos",
+					new MemoryDBSaveException("La clase articulo o key son null"));
+		} else if (valor instanceof Articulo) {
+			throw new MemoryDBCRUDException("Error al Guardar en la Base de Datos",
+					new MemoryDBSaveException("El objeto ingresado no es una instancia de Articulo"));
 		}
+
+	}
+
+	public static void update(Long key, Articulo newValue) throws MemoryDBCRUDException {
+		if (key != null && newValue != null && newValue instanceof Articulo) {
+			InMemoryDB.db.replace(key, newValue);
+		} else if (key == null || newValue == null) {
+			throw new MemoryDBCRUDException("Error al Modificar en la Base de Datos",
+					new MemoryDBUpdateException("La clase articulo o key son null"));
+		} else if (newValue instanceof Articulo) {
+			throw new MemoryDBCRUDException("Error al Modificar en la Base de Datos",
+					new MemoryDBUpdateException("El objeto ingresado no es una instancia de Articulo"));
+		}
+	}
+
+	public static void delete(Long key) throws MemoryDBCRUDException {
+		if (key != null) {
+			InMemoryDB.db.remove(key);
+		} else if (key == null) {
+			throw new MemoryDBCRUDException("Error al Borrar en la Base de Datos",
+					new MemoryDBUpdateException("La key es null"));
+		}
+	}
+
+	public static Articulo getById(Long key) throws MemoryDBCRUDException {
+		Articulo value = null;
+
+		if (key != null) {
+			if (InMemoryDB.db.containsKey(key)) {
+				value = InMemoryDB.db.get(key);
+			}
+		}else if(key == null){
+			throw new MemoryDBCRUDException("Error al Buscar en la Base de Datos",
+					new MemoryDBModificarException("La key es null"));
+		}
+		
 		return value;
 	}
 
-	public static void listAll() throws MemoryDBException{
-		System.out.println(InMemoryDB.db.values());		
+	public static void listAll() {
+		System.out.println(InMemoryDB.db.values());
+
+		
 	}
 }
